backend/images/analysis/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--no-binary shapely

branca==0.4.0             # via folium
cartopy==0.18.0           # via -r requirements.in
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via networkx
folium==0.10.1            # via -r corona-analysis/requirements.txt
idna==2.9                 # via requests
jinja2==2.11.1            # via branca, folium
kiwisolver==1.2.0         # via matplotlib
llvmlite==0.31.0          # via numba
markupsafe==1.1.1         # via jinja2
matplotlib==3.2.1         # via -r corona-analysis/requirements.txt
more-itertools==8.2.0     # via -r corona-analysis/requirements.txt
networkx==2.4             # via -r corona-analysis/requirements.txt
numba==0.48.0             # via -r corona-analysis/requirements.txt
numpy==1.18.2             # via cartopy, folium, matplotlib, numba, pandas, pykdtree, scipy, wquantiles
pandas==1.0.3             # via -r corona-analysis/requirements.txt
pillow==7.1.2             # via -r requirements.in
pyjwt==1.7.1              # via -r requirements.in
pykdtree==1.3.1           # via -r requirements.in
pyodbc==4.0.30            # via -r corona-analysis/requirements.txt
pyparsing==2.4.7          # via matplotlib
pyshp==2.1.0              # via cartopy
python-dateutil==2.8.1    # via matplotlib, pandas
pytz==2019.3              # via pandas
redis==3.4.1              # via -r requirements.in, redlock-py
redlock-py==1.0.8         # via -r requirements.in
requests==2.23.0          # via -r corona-analysis/requirements.txt, folium
scipy==1.4.1              # via -r requirements.in
shapely==1.7.0 ; sys_platform != "win32"  # via -r corona-analysis/requirements.txt, cartopy
six==1.14.0               # via branca, cartopy, cycler, python-dateutil
tornado==6.0.4            # via -r requirements.in
tqdm==4.45.0              # via -r corona-analysis/requirements.txt
urllib3==1.25.8           # via requests
wquantiles==0.5           # via -r corona-analysis/requirements.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
